JWTAuthentication was implemented.

+++ authentication/JWTAuth/Sources/App/Controllers/ParksController.swift
@@ -1,17 +1,25 @@
 import Foundation
 import Hummingbird
+import HummingbirdAuth
+import JWTKit
 import Logging
 import OracleNIO
 
-struct ParksController<Context: RequestContext> {
+struct ParksController {
   let client: OracleClient
   let logger: Logger
+  let jwtKeyCollection: JWTKeyCollection
+  let kid: JWKIdentifier
+
+  typealias Context = AppRequestContext
 
   func addRoutes(to group: RouterGroup<Context>) {
     group
-      .post(use: create)
+
       .get(use: index)
       .get(":id", use: show)
+      .add(middleware: JWTAuthenticator(jwtKeyCollection: jwtKeyCollection))
+      .post(use: create)
       .patch(":id", use: update)
       .delete(":id", use: delete)
   }
@@ -20,14 +28,22 @@ struct ParksController<Context: RequestContext> {
   /// Creates a new park with `coordinates` and `details`
   @Sendable
   func create(_ request: Request, context: Context) async throws -> HTTPResponse.Status {
-    let park = try await request.decode(as: Park.self, context: context)
+    let park = try await request.decode(as: NewPark.self, context: context)
 
     let detailsJSON = OracleJSON(park.details)
+    guard let userId = context.identity?.id?.uuidString.replacingOccurrences(of: "-", with: "") else {
+      return .unauthorized
+    }
 
-    let query: OracleStatement = try "INSERT INTO parks (coordinates, details) VALUES (SDO_GEOMETRY(\(park.coordinates.latitude), \(park.coordinates.longitude)), \(detailsJSON))"
+    let query: OracleStatement = try "INSERT INTO parks (coordinates, details, user_id) VALUES (SDO_GEOMETRY(\(park.coordinates.latitude), \(park.coordinates.longitude)), \(detailsJSON), \(userId))"
 
     _ = try await client.withConnection { conn in
-      try await conn.execute(query, logger: logger)
+      do {
+        try await conn.execute(query, logger: logger)
+      }
+      catch {
+        print(String(reflecting: error))
+      }
     }
 
     return .created
@@ -46,18 +62,20 @@ struct ParksController<Context: RequestContext> {
           id,
            p.coordinates.SDO_POINT.X AS latitude,
            p.coordinates.SDO_POINT.Y AS longitude,
-           p.details
+           p.details,
+           p.user_id
         FROM
           parks p
         """
       )
 
-      for try await (id, latitude, longitude, details) in stream.decode((UUID, Float, Float, OracleJSON<Park.Details>).self) {
+      for try await (id, latitude, longitude, details, user_id) in stream.decode((UUID, Float, Float, OracleJSON<Park.Details>, UUID).self) {
         parks.append(
           .init(
             id: id,
             coordinates: Park.Coordinates.init(latitude: latitude, longitude: longitude),
-            details: Park.Details.init(name: details.value.name)
+            details: Park.Details.init(name: details.value.name),
+            userId: user_id
           )
         )
       }
@@ -79,18 +97,20 @@ struct ParksController<Context: RequestContext> {
           id,
            p.coordinates.SDO_POINT.X AS latitude,
            p.coordinates.SDO_POINT.Y AS longitude,
-           p.details
+           p.details,
+           p.user_id
         FROM
           parks p
         WHERE id = HEXTORAW(\(guid))
         """
       )
 
-      for try await (id, latitude, longitude, details) in stream.decode((UUID, Float, Float, OracleJSON<Park.Details>).self) {
+      for try await (id, latitude, longitude, details, user_id) in stream.decode((UUID, Float, Float, OracleJSON<Park.Details>, UUID).self) {
         return Park(
           id: id,
           coordinates: Park.Coordinates.init(latitude: latitude, longitude: longitude),
-          details: Park.Details.init(name: details.value.name)
+          details: Park.Details.init(name: details.value.name),
+          userId: user_id
         )
       }
 
@@ -102,14 +122,64 @@ struct ParksController<Context: RequestContext> {
   /// Updates a single park with id
   @Sendable
   func update(_ request: Request, context: Context) async throws -> HTTPResponse.Status {
+    let userInput = try await request.decode(as: UpdatePark.self, context: context)
+    // User
+    guard let userId = context.identity?.id?.uuidString.replacingOccurrences(of: "-", with: "") else {
+      return .badRequest
+    }
+
+    // Park
+    var originalPark: Park?
     let id = try context.parameters.require("id", as: String.self)
     let guid = id.replacingOccurrences(of: "-", with: "")
-    let park = try await request.decode(as: Park.self, context: context)
-    let detailsJSON = OracleJSON(park.details)
+
+    // Get original park
+    _ = try await client.withConnection { conn in
+      let stream = try await conn.execute(
+        """
+        SELECT
+          id,
+           p.coordinates.SDO_POINT.X AS latitude,
+           p.coordinates.SDO_POINT.Y AS longitude,
+           p.details,
+           p.user_id
+        FROM
+          parks p
+        WHERE id = HEXTORAW(\(guid))
+        AND user_id = HEXTORAW(\(userId))
+        """
+      )
+
+      for try await (id, latitude, longitude, details, user_id) in stream.decode((UUID, Float, Float, OracleJSON<Park.Details>, UUID).self) {
+        originalPark = Park(
+          id: id,
+          coordinates: Park.Coordinates.init(
+            latitude: latitude,
+            longitude: longitude
+          ),
+          details: Park.Details.init(name: details.value.name),
+          userId: user_id
+        )
+      }
+    }
+
+    // Check if the originalPark in the database
+    guard let originalPark = originalPark else {
+      throw HTTPError(.notFound, message: "Park was not found")
+    }
+
+    // Check if there is an update, if not use the original values
+    let latitude = userInput.coordinates?.latitude ?? originalPark.coordinates.latitude
+    let longitude = userInput.coordinates?.longitude ?? originalPark.coordinates.longitude
+    let details = Park.Details(
+      name: userInput.details?.name ?? originalPark.details.name
+    )
+
+    let detailsJSON = OracleJSON(details)
 
     let query: OracleStatement = try """
     UPDATE parks
-    SET coordinates = SDO_GEOMETRY(\(park.coordinates.latitude), \(park.coordinates.longitude)),
+    SET coordinates = SDO_GEOMETRY(\(latitude), \(longitude)),
         details = \(detailsJSON)
     WHERE id = HEXTORAW(\(guid))
     """
@@ -120,7 +190,6 @@ struct ParksController<Context: RequestContext> {
       if updatedRows == 0 {
         return .notFound
       }
-
       return .ok
     }
   }
@@ -129,6 +198,12 @@ struct ParksController<Context: RequestContext> {
   /// Deletes park with id
   @Sendable
   func delete(_: Request, context: Context) async throws -> HTTPResponse.Status {
+    // User
+    guard let userId = context.identity?.id?.uuidString.replacingOccurrences(of: "-", with: "") else {
+      return .badRequest
+    }
+
+    // Park
     let id = try context.parameters.require("id", as: String.self)
     let guid = id.replacingOccurrences(of: "-", with: "")
 
@@ -137,14 +212,15 @@ struct ParksController<Context: RequestContext> {
         """
         DELETE FROM parks
         WHERE id = HEXTORAW(\(guid))
+        AND user_id = HEXTORAW(\(userId))
         """,
         logger: logger
       )
       let deletedRows = try await stream.affectedRows
       if deletedRows == 0 {
-        return .notFound
+        throw HTTPError(.badRequest, message: "The park either doesn't exist or you have no persmission to delete it.")
       }
-      return .ok
+      return .noContent
     }
   }
 }
